добавлены Cloud) unit CRUD -tests для users, accounts

+++ src/test/java/com/minibank/demo/repository/UserRepositoryTest.java
@@ -1,35 +1,44 @@
 package com.minibank.demo.repository;
 
 import com.minibank.demo.model.User;
-
-import org.junit.jupiter.api.Order;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.assertj.core.api.Assertions;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.test.annotation.Rollback;
+
 
 import java.util.List;
 import java.util.Optional;
 
 
 @DataJpaTest
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 class UserRepositoryTest {
     @Autowired
     private UserRepository userRepository;
 
+    User user;
+    @BeforeEach
+    void setUp() {
+        user = new User(1L,"Ivan","Ivanov", "Ivanovich","ivv","ivv@mail.com","xxx");
+        userRepository.save(user);
+    }
 
+    @AfterEach
+    void tearDown() {
+        user = null;
+        userRepository.deleteAll();
+    }
     @Test
-    @Order(0)
+
     public void findById() {
         User user = userRepository.findById(1L).get();
         Assertions.assertThat(user.getId()).isEqualTo(1L);
     }
 
     @Test
-    @Order(1)
+
     public void getListOfUserTest(){
 
         List<User> users = userRepository.findAll();
@@ -39,37 +48,41 @@ class UserRepositoryTest {
     }
 
     @Test
-    @Order(2)
+
     public void updateUserTest(){
 
-        User user = userRepository.findById(1L).get();
+        if (userRepository.findById(1L).isPresent()) {
 
-        user.setEmail("ram@gmail.com");
+            User user = userRepository.findById(1L).get();
 
-        User userUpdated =  userRepository.save(user);
+            user.setEmail("ram@gmail.com");
 
-        Assertions.assertThat(userUpdated.getEmail()).isEqualTo("ram@gmail.com");
+            User userUpdated = userRepository.save(user);
 
+            Assertions.assertThat(userUpdated.getEmail()).isEqualTo("ram@gmail.com");
+        }
     }
 
     @Test
-    @Order(3)
-    public void deleteUserTest(){
 
-        User user = userRepository.findById(1L).get();
+    public void deleteUserTest() {
 
-        userRepository.delete(user);
+        if (userRepository.findById(1L).isPresent()) {
+            User user = userRepository.findById(1L).get();
 
-        //userRepository.deleteById(1L);
+            userRepository.delete(user);
 
-        User user1 = null;
+            //userRepository.deleteById(1L);
 
-        Optional<User> optionalUser = userRepository.findByUserName("iw");
+            User user1 = null;
 
-        if(optionalUser.isPresent()){
-            user1 = optionalUser.get();
+            Optional<User> optionalUser = userRepository.findByUserName("iw");
+
+            if (optionalUser.isPresent()) {
+                user1 = optionalUser.get();
+            }
+
+            Assertions.assertThat(user1).isNull();
         }
-
-        Assertions.assertThat(user1).isNull();
     }
 }
